Bees_docker/requirements.txt
requests
pandas
google-cloud-bigquery
google-cloud-storage
pyspark
gcsfs
pyarrow